+++ packages/core/src/node/messaging/messaging-service.ts
@@ -15,33 +15,21 @@
 // *****************************************************************************
 
 import { Socket } from 'socket.io';
-import { MessageConnection } from 'vscode-ws-jsonrpc';
-import { IConnection } from 'vscode-ws-jsonrpc/lib/server/connection';
-import { WebSocketChannel } from '../../common/messaging/web-socket-channel';
+import { Channel } from '../../common/message-rpc/channel';
 
 export interface MessagingService {
-    /**
-     * Accept a JSON-RPC connection on the given path.
-     * A path supports the route syntax: https://github.com/rcs/route-parser#what-can-i-use-in-my-routes.
-     */
-    listen(path: string, callback: (params: MessagingService.PathParams, connection: MessageConnection) => void): void;
-    /**
-     * Accept a raw JSON-RPC connection on the given path.
-     * A path supports the route syntax: https://github.com/rcs/route-parser#what-can-i-use-in-my-routes.
-     */
-    forward(path: string, callback: (params: MessagingService.PathParams, connection: IConnection) => void): void;
     /**
      * Accept a web socket channel on the given path.
      * A path supports the route syntax: https://github.com/rcs/route-parser#what-can-i-use-in-my-routes.
      */
-    wsChannel(path: string, callback: (params: MessagingService.PathParams, socket: WebSocketChannel) => void): void;
+    wsChannel(path: string, callback: (params: MessagingService.PathParams, channel: Channel) => void): void;
     /**
      * Accept a web socket connection on the given path.
      * A path supports the route syntax: https://github.com/rcs/route-parser#what-can-i-use-in-my-routes.
      *
      * #### Important
-     * Prefer JSON-RPC connections or web socket channels over web sockets. Clients can handle only limited amount of web sockets
-     * and excessive amount can cause performance degradation. All JSON-RPC connections and web socket channels share the single web socket connection.
+     * Prefer using web socket channels over establishing new web socket connection. Clients can handle only limited amount of web sockets
+     * and excessive amount can cause performance degradation. All web socket channels share a single web socket connection.
      */
     ws(path: string, callback: (params: MessagingService.PathParams, socket: Socket) => void): void;
 }
@@ -54,20 +42,5 @@ export namespace MessagingService {
     export const Contribution = Symbol('MessagingService.Contribution');
     export interface Contribution {
         configure(service: MessagingService): void;
-    }
-}
-
-export interface WebSocketChannelConnection extends IConnection {
-    channel: WebSocketChannel;
-}
-export namespace WebSocketChannelConnection {
-    export function is(connection: IConnection): connection is WebSocketChannelConnection {
-        return (connection as WebSocketChannelConnection).channel instanceof WebSocketChannel;
-    }
-
-    export function create(connection: IConnection, channel: WebSocketChannel): WebSocketChannelConnection {
-        const result = connection as WebSocketChannelConnection;
-        result.channel = channel;
-        return result;
     }
 }