+++ packages/terminal/src/node/terminal-backend-contribution.slow-spec.ts
@@ -20,7 +20,7 @@ import { IShellTerminalServer } from '../common/shell-terminal-protocol';
 import * as http from 'http';
 import * as https from 'https';
 import { terminalsPath } from '../common/terminal-protocol';
-import { TestWebSocketChannel } from '@theia/core/lib/node/messaging/test/test-web-socket-channel';
+import { TestWebSocketChannelSetup } from '@theia/core/lib/node/messaging/test/test-web-socket-channel';
 
 describe('Terminal Backend Contribution', function (): void {
 
@@ -45,13 +45,19 @@ describe('Terminal Backend Contribution', function (): void {
     it('is data received from the terminal ws server', async () => {
         const terminalId = await shellTerminalServer.create({});
         await new Promise<void>((resolve, reject) => {
-            const channel = new TestWebSocketChannel({ server, path: `${terminalsPath}/${terminalId}` });
+            const path = `${terminalsPath}/${terminalId}`;
+            const { channel, multiplexer } = new TestWebSocketChannelSetup({ server, path });
             channel.onError(reject);
-            channel.onClose((code, reason) => reject(new Error(`channel is closed with '${code}' code and '${reason}' reason`)));
-            channel.onOpen(() => {
-                resolve();
-                channel.close();
+            channel.onClose(event => reject(new Error(`channel is closed with '${event.code}' code and '${event.reason}' reason}`)));
+
+            multiplexer.onDidOpenChannel(event => {
+                if (event.id === path) {
+                    resolve();
+                    channel.close();
+                }
             });
+
         });
     });
+
 });